checking id for being in the company
